Move functions into new jsonflib package
Extract the useful `Query` and `Highlight` functions into a new
`jsonflib` package, so that other Golang programs can import
`github.com/miolini/jsonf/jsonflib` and use these functions.

Example usage:

```go
import (
  ...
  "github.com/miolini/jsonf/jsonflib"
  ...
)
...
    jsonRaw, err := jsonflib.Highlight(
      jsonRaw,
      jsonflib.HighlightFlags{Colorize: true, Verbose: false, Debug: false},
    )
    if err != nil {
      return err
    }
```
